+++ streamer.py
@@ -8,7 +8,7 @@ from autobahn.twisted.websocket import WebSocketClientFactory, \
     WebSocketClientProtocol, \
     connectWS
 
-from util import MSG_TYPE_NEW, MSG_TYPE_RESUME, MSG_TYPE_END, \
+from util.constants import MSG_TYPE_NEW, MSG_TYPE_RESUME, MSG_TYPE_END, \
         MSG_TYPE_BUZZING_REQUEST, MSG_TYPE_BUZZING_ANSWER, \
         MSG_TYPE_BUZZING_GREEN, MSG_TYPE_BUZZING_RED
 ''' Streaming interface for updating questions in game. Also handles
